refactor: Move GtfsFeedId to o.o.gtfs.graphbuilder package
