+++ application/src/test/java/org/opentripplanner/graph_builder/module/GtfsFeedIdTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.opentripplanner.gtfs.graphbuilder.GtfsFeedId;
 
 class GtfsFeedIdTest {
