+++ application/src/test/java/org/opentripplanner/gtfs/GtfsContext.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.gtfs;
 
-import org.opentripplanner.graph_builder.module.GtfsFeedId;
+import org.opentripplanner.gtfs.graphbuilder.GtfsFeedId;
 import org.opentripplanner.model.OtpTransitService;
 import org.opentripplanner.model.calendar.CalendarServiceData;
